backend/requirements.txt
aiodns==3.2.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.4.0
app==0.0.1
astroid==3.2.4
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
blis==0.7.11
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
confection==0.1.5
contourpy==1.2.1
contractions==0.1.73
cycler==0.12.1
cymem==2.0.8
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
docopt==0.6.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
executing==2.0.1
fastapi==0.112.1
fastjsonschema==2.20.0
Flask==3.0.3
fonttools==4.53.1
frozenlist==1.4.1
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httplib2==0.22.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.8
iniconfig==2.0.0
ipython==8.12.3
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonify==0.5
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
Mako==1.3.5
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
multidict==6.0.5
murmurhash==1.0.10
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nltk==3.9.1
numpy==2.1.0
oauthlib==3.2.2
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.2.2
pluggy==1.5.0
preshed==3.0.9
priority==2.0.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.27.3
psutil==6.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pure_eval==0.2.3
pyahocorasick==2.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycares==4.4.0
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
Pygments==2.18.0
pylint==3.2.6
pyparsing==3.1.2
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
Quart==0.19.6
ratelimit==2.2.1
referencing==0.35.1
regex==2024.7.24
reportlab==4.2.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
seaborn==0.13.2
setuptools==73.0.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
soupsieve==2.6
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.32
srsly==2.4.8
stack-data==0.6.3
starlette==0.38.2
tabulate==0.9.0
textsearch==0.0.24
thinc==8.2.4
tinycss2==1.3.0
tk==0.1.0
tomlkit==0.13.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typer==0.12.4
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.6
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
Werkzeug==3.0.4
wrapt==1.16.0
wsproto==1.2.0
yarg==0.1.9
yarl==1.9.4
